Switch to IMAP
Switched to Gmail IMAP via App Logins to keep a longer session and authenticate without the refresh token expiring.

Verification of the app and implementing oauth2 requires further local endpoint listeners to help recieve the callback which may be considered down the line but not at the moment alongside privacy policies.